Rename contact list namespace
