+++ test/Abp.TestBase.SampleApplication.Tests/ContactLists/ContactList_MultiTenancy_Tests.cs
@@ -2,7 +2,7 @@
 using Abp.Configuration.Startup;
 using Abp.Domain.Repositories;
 using Abp.Domain.Uow;
-using Abp.TestBase.SampleApplication.ContacLists;
+using Abp.TestBase.SampleApplication.ContactLists;
 using Shouldly;
 using Xunit;
