+++ test/Abp.TestBase.SampleApplication.Tests/EntityFramework/GraphDiff/GraphDiffExtensions_Tests.cs
@@ -2,7 +2,7 @@
 using Abp.Domain.Repositories;
 using Abp.Domain.Uow;
 using Abp.EntityFramework.GraphDiff.Extensions;
-using Abp.TestBase.SampleApplication.ContacLists;
+using Abp.TestBase.SampleApplication.ContactLists;
 using Abp.TestBase.SampleApplication.People;
 using Shouldly;
 using Xunit;
