+++ test/Abp.TestBase.SampleApplication.Tests/People/PersonAppService_Tests.cs
@@ -5,7 +5,7 @@ using Abp.Application.Services.Dto;
 using Abp.Authorization;
 using Abp.Domain.Uow;
 using Abp.Runtime.Validation;
-using Abp.TestBase.SampleApplication.ContacLists;
+using Abp.TestBase.SampleApplication.ContactLists;
 using Abp.TestBase.SampleApplication.People;
 using Abp.TestBase.SampleApplication.People.Dto;
 using Castle.MicroKernel.Registration;
